CHG: Examples and read

+++ Laboratorio 1/include/image.h
@@ -1,10 +1,18 @@
 # ifndef IMAGE_H
 # define IMAGE_H
 
+# include <stdio.h>
+# include <stdlib.h>
+# include <string.h>
+# include <ctype.h>
+# include <unistd.h>
+# include <pmmintrin.h>
+# include <immintrin.h>
+
 typedef struct image {
 	int size;
 	char* filename;
-	int** content;
+	float** content;
 } image;
 
 image* readImage(FILE* file, int size);
